Fix (most) compiler warnings

+++ c++/triqs/utility/time_pt.hpp
@@ -56,7 +56,7 @@ namespace triqs {
       }
 
       public:
-      time_pt(uint64_t n_, double beta_) : beta(beta_), n(n_), val(beta_ * (double(n_) / Nmax)) {}
+      time_pt(uint64_t n_, double beta_) : beta(beta_), n(n_), val(beta_ * (static_cast<double>(n_) / static_cast<double>(Nmax))) {}
 
       /// Forbidden
       time_pt &operator=(double v) = delete;
@@ -165,7 +165,7 @@ namespace triqs {
       /// ??
       time_pt make_time_pt(double x) const {
         EXPECTS(0 <= x && x <= beta);
-        uint64_t n = time_pt::Nmax * std::min(1.0, (x / beta));
+        uint64_t n = static_cast<double>(time_pt::Nmax) * std::min(1.0, (x / beta));
         return time_pt(n, beta);
       }
     };
